fiexed event id column in attendees

+++ database/migrations/2022_12_18_193705_create_attendees_table.php
@@ -15,9 +15,9 @@ class CreateAttendeesTable extends Migration
     {
         Schema::create('attendees', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('order_id')->index();
-            $table->unsignedInteger('event_id')->index();
-            $table->unsignedInteger('ticket_id')->index();
+            $table->unsignedBigInteger('order_id')->index();
+            $table->unsignedBigInteger('event_id')->index();
+            $table->unsignedBigInteger('ticket_id')->index();
 
             $table->string('fullname');
             $table->string('email');
@@ -25,7 +25,6 @@ class CreateAttendeesTable extends Migration
             $table->string('reference', 20);
 
             $table->dateTime('check_time')->nullable();
-
 
             $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
             $table->foreign('ticket_id')->references('id')->on('tickets')->onDelete('cascade');
